+++ src/lib/fs.test.ts
@@ -1,10 +1,10 @@
 import {test} from 'uvu';
 import * as assert from 'uvu/assert';
 
-import {exists} from './exists.js';
+import {exists} from './fs.js';
 
 test('file exists', async () => {
-	assert.ok(await exists('./src/lib/exists.ts'));
+	assert.ok(await exists('./src/lib/fs.ts'));
 });
 
 test('file does not exist', async () => {